Fixed Task API and Completed Task Creation
There was some issues with the Task API because I added an additional /tasks to the addresses, so the requests weren't going through properly.

Updated task model to include a due date, and updated the form to include the status of the task.

TODO: Clear out all of the unnecessary console.log() statements.